Rename MultiBadge to BadgeSelected
