+++ packages/ui-components/src/components/form/select/BadgeSelected.vue
